[Keyboard] Organized versions of mikeneko65 (#16985)
