+++ keyboards/mikeneko65/config.h
@@ -21,14 +21,14 @@ along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
 /* USB Device descriptor parameter */
 #define VENDOR_ID    0x7934
-#define PRODUCT_ID   0x6D53
+#define PRODUCT_ID   0x6D54
 #define DEVICE_VER   0x0001
 #define MANUFACTURER takishim
-#define PRODUCT      Mikeneko 65 V3
+#define PRODUCT      Mikeneko 65
 
 /* key matrix size */
 #define MATRIX_ROWS 5
-#define MATRIX_COLS 16
+#define MATRIX_COLS 15
 
 /*
  * Keyboard Matrix Assignments
@@ -40,8 +40,8 @@ along with this program.  If not, see <http://www.gnu.org/licenses/>.
  *                  ROW2COL = ROW = Anode (+), COL = Cathode (-, marked on diode)
  *
  */
-#define MATRIX_ROW_PINS { E6, B7, F7, F4, F5 }
-#define MATRIX_COL_PINS { F6, B0, F1, C7, C6, B6, B5, B4, D7, D6, D4, D5, D3, D2, D1, D0 }
+#define MATRIX_ROW_PINS { D0, D1, D2, D3, C7 }
+#define MATRIX_COL_PINS { F7, F6, F5, F4, F1, F0, E6, B0, B7, D4, D6, D7, B6, B5, B4 }
 #define UNUSED_PINS
 
 /* COL2ROW, ROW2COL */
