+++ keyboards/mikeneko65/info.json
@@ -1,5 +1,5 @@
 {
-    "keyboard_name": "Mikeneko 65 V4",
+    "keyboard_name": "Mikeneko 65",
     "url": "https://github.com/takishim/mikeneko65",
     "maintainer": "takishim",
     "layouts": {
