+++ keyboards/mikeneko65/keymaps/default/readme.md
@@ -1,0 +1,1 @@
+# The default keymap for Mikeneko 65
