+++ keyboards/mikeneko65/mikeneko65.c
@@ -15,4 +15,4 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-#include "rev4.h"
+#include "mikeneko65.h"
