+++ keyboards/mikeneko65/readme.md
@@ -1,18 +1,18 @@
-# Mikeneko 65 V3
+# Mikeneko 65
 
 A simple 65% keyboard
 
 * Keyboard Maintainer: [takishim](https://github.com/takishim)
-* Hardware Supported: Mikeneko 65 V4
-* Hardware Availability: [Open source on GitHub](https://github.com/takishim/mikeneko65/tree/v3)
+* Hardware Supported: Mikeneko 65
+* Hardware Availability: [Open source on GitHub](https://github.com/takishim/mikeneko65/)
 
 Make example for this keyboard (after setting up your build environment):
 
-    make mikeneko65/rev3:default
+    make mikeneko65:default
 
 Flashing example for this keyboard:
 
-    make mikeneko65/rev3:default:flash
+    make mikeneko65:default:flash
 
 Enter the bootloader in 3 ways:
 
@@ -20,4 +20,4 @@ Enter the bootloader in 3 ways:
 * **Physical reset button**: Briefly press the button on the back of the PCB - some may have pads you must short instead
 * **Keycode in layout**: Press the key mapped to `RESET` if it is available
 
-See the [build environment setup](https://docs.qmk.fm/#/getting_started_build_tools) and the [make instructions](https://docs.qmk.fm/#/getting_started_make_guide) for more information. Brand new to QMK? Start with our [Complete Newbs Guide](https://docs.qmk.fm/#/newbs).
+ぶSee the [build environment setup](https://docs.qmk.fm/#/getting_started_build_tools) and the [make instructions](https://docs.qmk.fm/#/getting_started_make_guide) for more information. Brand new to QMK? Start with our [Complete Newbs Guide](https://docs.qmk.fm/#/newbs).
