--- keyboards/mikeneko65/rev4/keymaps/default/readme.md
@@ -1,1 +1,0 @@
-# The default keymap for Mikeneko 65 V4
